Try to fix slowdown
